+++ packages/atomic/src/components/common/atomic-focus-trap.tsx
@@ -1,34 +1,11 @@
 import {Component, Element, Listen, Prop, Watch} from '@stencil/core';
 import {getFirstFocusableDescendant} from '../../utils/accessibility-utils';
-import {defer, getFocusedElement} from '../../utils/utils';
-
-function getParent(element: Element | ShadowRoot) {
-  if (element.parentNode) {
-    return element.parentNode as Element | ShadowRoot;
-  }
-  if (element instanceof ShadowRoot) {
-    return element.host;
-  }
-  return null;
-}
-
-function contains(
-  ancestor: Element | ShadowRoot,
-  element: Element | ShadowRoot
-): boolean {
-  if (element === ancestor) {
-    return true;
-  }
-  if (
-    element instanceof HTMLElement &&
-    element.assignedSlot &&
-    contains(ancestor, element.assignedSlot)
-  ) {
-    return true;
-  }
-  const parent = getParent(element);
-  return parent === null ? false : contains(ancestor, parent);
-}
+import {
+  isAncestorOf,
+  defer,
+  getFocusedElement,
+  getParent,
+} from '../../utils/utils';
 
 /**
  * @internal
@@ -48,6 +25,12 @@ export class AtomicFocusTrap {
    * The container to hide from the tabindex and accessibility DOM when the focus trap is inactive.
    */
   @Prop({mutable: true}) container?: HTMLElement;
+
+  /**
+   * Whether the element should be hidden from screen readers & not interactive with the tab, when not active.
+   */
+  @Prop() shouldHideSelf = true;
+
   private readonly hiddenElements: Element[] = [];
 
   hide(element: Element) {
@@ -74,7 +57,10 @@ export class AtomicFocusTrap {
       if (sibling === element) {
         return;
       }
-      if (sibling.assignedSlot && contains(this.host, sibling.assignedSlot)) {
+      if (
+        sibling.assignedSlot &&
+        isAncestorOf(this.host, sibling.assignedSlot)
+      ) {
         return;
       }
       this.hide(sibling);
@@ -90,8 +76,10 @@ export class AtomicFocusTrap {
   }
 
   hideSelf() {
-    this.parentToHide.setAttribute('aria-hidden', 'true');
-    this.parentToHide.setAttribute('tabindex', '-1');
+    if (this.shouldHideSelf) {
+      this.parentToHide.setAttribute('aria-hidden', 'true');
+      this.parentToHide.setAttribute('tabindex', '-1');
+    }
   }
 
   async onDeactivated(isInitialLoad: boolean) {
@@ -130,7 +118,7 @@ export class AtomicFocusTrap {
 
     const focusedElement = getFocusedElement();
 
-    if (focusedElement && contains(this.host, focusedElement)) {
+    if (focusedElement && isAncestorOf(this.host, focusedElement)) {
       return;
     }