+++ packages/atomic/src/components/common/facets/numeric-facet-common.tsx
@@ -1,4 +1,12 @@
 import {Schema, StringValue} from '@coveo/bueno';
+import {VNode, h} from '@stencil/core';
+import {FocusTargetController} from '../../../utils/accessibility-utils';
+import {getFieldValueCaption} from '../../../utils/field-utils';
+import {randomID} from '../../../utils/utils';
+import {initializePopover} from '../../search/facets/atomic-popover/popover-type';
+import {NumberFormatter} from '../formats/format-common';
+import {Hidden} from '../hidden';
+import {AnyBindings} from '../interface/bindings';
 import {
   FacetConditionsManager,
   NumericFacet,
@@ -7,23 +15,7 @@ import {
   NumericRangeOptions,
   NumericRangeRequest,
   SearchStatusState,
-} from '@coveo/headless';
-import {VNode, h} from '@stencil/core';
-import {FocusTargetController} from '../../../utils/accessibility-utils';
-import {getFieldValueCaption} from '../../../utils/field-utils';
-import {
-  InsightFacetConditionsManager,
-  InsightNumericFacet,
-  InsightNumericFacetValue,
-  InsightNumericFilter,
-  InsightNumericRangeRequest,
-  InsightSearchStatusState,
-} from '../../insight';
-import {InsightBindings} from '../../insight/atomic-insight-interface/atomic-insight-interface';
-import {Bindings} from '../../search/atomic-search-interface/atomic-search-interface';
-import {initializePopover} from '../../search/facets/atomic-popover/popover-type';
-import {NumberFormatter} from '../formats/format-common';
-import {Hidden} from '../hidden';
+} from '../types';
 import {
   shouldDisplayInputForFacetRange,
   validateDependsOn,
@@ -38,10 +30,6 @@ import {FacetValueLabelHighlight} from './facet-value-label-highlight/facet-valu
 import {FacetValueLink} from './facet-value-link/facet-value-link';
 import {FacetValuesGroup} from './facet-values-group/facet-values-group';
 
-export interface InsightNumericRangeWithLabel
-  extends InsightNumericRangeRequest {
-  label?: string;
-}
 export interface NumericRangeWithLabel extends NumericRangeRequest {
   label?: string;
 }
@@ -49,8 +37,9 @@ export interface NumericRangeWithLabel extends NumericRangeRequest {
 export type NumericFacetDisplayValues = 'checkbox' | 'link';
 
 interface NumericFacetCommonOptions {
+  facetId?: string;
   host: HTMLElement;
-  bindings: Bindings | InsightBindings;
+  bindings: AnyBindings;
   label: string;
   field: string;
   headingLevel: number;
@@ -60,19 +49,15 @@ interface NumericFacetCommonOptions {
   numberOfValues: number;
   setFacetId(id: string): string;
   setManualRanges(
-    manualRanges: InsightNumericRangeWithLabel[] | NumericRangeWithLabel[]
-  ): InsightNumericRangeWithLabel[] | NumericRangeWithLabel[];
+    manualRanges: NumericRangeWithLabel[]
+  ): NumericRangeWithLabel[];
   getFormatter(): NumberFormatter;
-  getSearchStatusState(): InsightSearchStatusState | SearchStatusState;
-  buildDependenciesManager():
-    | InsightFacetConditionsManager
-    | FacetConditionsManager;
-  buildNumericRange(
-    config: NumericRangeOptions
-  ): InsightNumericRangeRequest | NumericRangeRequest;
-  initializeFacetForInput(): InsightNumericFacet | NumericFacet;
-  initializeFacetForRange(): InsightNumericFacet | NumericFacet;
-  initializeFilter(): InsightNumericFilter | NumericFilter;
+  getSearchStatusState(): SearchStatusState;
+  buildDependenciesManager(): FacetConditionsManager;
+  buildNumericRange(config: NumericRangeOptions): NumericRangeRequest;
+  initializeFacetForInput(): NumericFacet;
+  initializeFacetForRange(): NumericFacet;
+  initializeFilter(): NumericFilter;
 }
 
 interface NumericFacetCommonRenderProps {
@@ -84,76 +69,61 @@ interface NumericFacetCommonRenderProps {
 }
 
 export class NumericFacetCommon {
-  private host: HTMLElement;
-  private bindings: Bindings | InsightBindings;
-  private label: string;
-  private field: string;
-  private headingLevel: number;
-  private filter?: InsightNumericFilter | NumericFilter;
-  private dependsOn: Record<string, string>;
-  private displayValuesAs: NumericFacetDisplayValues;
-  private withInput?: NumberInputType;
-  private numberOfValues: number;
-  private facetId?: string;
-  private manualRanges:
-    | InsightNumericRangeWithLabel[]
-    | NumericRangeWithLabel[] = [];
-  private facetForRange?: InsightNumericFacet | NumericFacet;
-  private facetForInput?: InsightNumericFacet | NumericFacet;
-  private getFormatter: () => NumberFormatter;
-  private getSearchStatusState: () =>
-    | InsightSearchStatusState
-    | SearchStatusState;
+  private facetId: string;
+  private filter?: NumericFilter;
+  private manualRanges: NumericRangeWithLabel[] = [];
+  private facetForRange?: NumericFacet;
+  private facetForInput?: NumericFacet;
 
-  private dependenciesManager:
-    | InsightFacetConditionsManager
-    | FacetConditionsManager;
+  private dependenciesManager: FacetConditionsManager;
 
-  constructor(props: NumericFacetCommonOptions) {
-    this.host = props.host;
-    this.bindings = props.bindings;
-    this.label = props.label;
-    this.field = props.field;
-    this.headingLevel = props.headingLevel;
-    this.dependsOn = props.dependsOn;
-    this.displayValuesAs = props.displayValuesAs;
-    this.withInput = props.withInput;
-    this.numberOfValues = props.numberOfValues;
-    this.getFormatter = props.getFormatter;
-    this.getSearchStatusState = props.getSearchStatusState;
+  constructor(private props: NumericFacetCommonOptions) {
     this.validateProps();
+    this.facetId = this.determineFacetId;
+    this.props.setFacetId(this.facetId);
 
     // Initialize two facets: One that is actually used to display values for end users, which only exists
     // if we need to display something to the end user (ie: numberOfValues > 0)
 
     // A second facet is initialized only to verify the results count. It is never used to display results to end user.
     // It serves as a way to determine if the input should be rendered or not, independent of the ranges (manual or automatic) configured in the component
-    if (this.numberOfValues > 0) {
-      this.manualRanges = props.setManualRanges(
-        Array.from(this.host.querySelectorAll('atomic-numeric-range')).map(
-          ({start, end, endInclusive, label}) => ({
-            ...props.buildNumericRange({start, end, endInclusive}),
-            label,
-          })
-        )
+    if (this.props.numberOfValues > 0) {
+      this.manualRanges = this.props.setManualRanges(
+        Array.from(
+          this.props.host.querySelectorAll('atomic-numeric-range')
+        ).map(({start, end, endInclusive, label}) => ({
+          ...this.props.buildNumericRange({start, end, endInclusive}),
+          label,
+        }))
       );
-      this.facetForRange = props.initializeFacetForRange();
-      this.facetId = props.setFacetId(this.facetForRange.state.facetId);
+      this.facetForRange = this.props.initializeFacetForRange();
     }
-    if (this.withInput) {
-      this.facetForInput = props.initializeFacetForInput();
-      this.filter = props.initializeFilter();
-      if (!this.facetId) {
-        this.facetId = this.filter.state.facetId;
-      }
+
+    if (this.props.withInput) {
+      this.facetForInput = this.props.initializeFacetForInput();
+      this.filter = this.props.initializeFilter();
     }
-    this.dependenciesManager = props.buildDependenciesManager();
+
+    this.dependenciesManager = this.props.buildDependenciesManager();
     this.registerFacetToStore();
   }
 
-  private get formatter() {
-    return this.getFormatter();
+  private get determineFacetId() {
+    if (this.props.facetId) {
+      return this.props.facetId;
+    }
+
+    if (this.props.bindings.store.get('numericFacets')[this.props.field]) {
+      return randomID(`${this.props.field}_`);
+    }
+
+    return this.props.field;
   }
+
+  private get formatter() {
+    return this.props.getFormatter();
+  }
+
   private get enabled() {
     return (
       this.facetForRange?.state.enabled ?? this.filter?.state.enabled ?? true
@@ -188,7 +158,7 @@ export class NumericFacetCommon {
   }
 
   private get searchStatusState() {
-    return this.getSearchStatusState();
+    return this.props.getSearchStatusState();
   }
 
   private get shouldRenderInput() {
@@ -196,14 +166,14 @@ export class NumericFacetCommon {
       hasInputRange: this.hasInputRange,
       searchStatusState: this.searchStatusState,
       facetValues: this.facetForInput?.state.values || [],
-      hasInput: !!this.withInput,
+      hasInput: !!this.props.withInput,
     });
   }
 
   private get shouldRenderValues() {
     return (
       !this.hasInputRange &&
-      this.numberOfValues > 0 &&
+      this.props.numberOfValues > 0 &&
       !!this.valuesToRender.length
     );
   }
@@ -212,14 +182,14 @@ export class NumericFacetCommon {
       displayValuesAs: new StringValue({constrainTo: ['checkbox', 'link']}),
       withInput: new StringValue({constrainTo: ['integer', 'decimal']}),
     }).validate({
-      displayValuesAs: this.displayValuesAs,
-      withInput: this.withInput,
+      displayValuesAs: this.props.displayValuesAs,
+      withInput: this.props.withInput,
     });
-    validateDependsOn(this.dependsOn);
+    validateDependsOn(this.props.dependsOn);
   }
 
   public disconnectedCallback() {
-    if (this.host.isConnected) {
+    if (this.props.host.isConnected) {
       return;
     }
     this.dependenciesManager?.stopWatching();
@@ -227,37 +197,39 @@ export class NumericFacetCommon {
 
   private registerFacetToStore() {
     const facetInfo: FacetInfo = {
-      label: this.label,
+      label: this.props.label,
       facetId: this.facetId!,
-      element: this.host,
+      element: this.props.host,
     };
 
-    this.bindings.store.registerFacet('numericFacets', {
+    this.props.bindings.store.registerFacet('numericFacets', {
       ...facetInfo,
       format: (value) => this.formatFacetValue(value),
     });
 
-    initializePopover(this.host, {
+    initializePopover(this.props.host, {
       ...facetInfo,
       hasValues: () => this.hasValues,
       numberOfSelectedValues: () => this.numberOfSelectedValues,
     });
 
     if (this.filter) {
-      this.bindings.store.state.numericFacets[this.filter.state.facetId] =
-        this.bindings.store.state.numericFacets[this.facetId!];
+      this.props.bindings.store.state.numericFacets[this.filter.state.facetId] =
+        this.props.bindings.store.state.numericFacets[this.facetId!];
     }
   }
 
-  private formatFacetValue(
-    facetValue: InsightNumericFacetValue | NumericFacetValue
-  ) {
+  private formatFacetValue(facetValue: NumericFacetValue) {
     const manualRangeLabel = this.manualRanges.find((range) =>
       this.areRangesEqual(range, facetValue)
     )?.label;
     return manualRangeLabel
-      ? getFieldValueCaption(this.field, manualRangeLabel, this.bindings.i18n)
-      : this.bindings.i18n.t('to', {
+      ? getFieldValueCaption(
+          this.props.field,
+          manualRangeLabel,
+          this.props.bindings.i18n
+        )
+      : this.props.bindings.i18n.t('to', {
           start: this.formatValue(facetValue.start),
           end: this.formatValue(facetValue.end),
         });
@@ -265,9 +237,12 @@ export class NumericFacetCommon {
 
   private formatValue(value: number) {
     try {
-      return this.formatter(value, this.bindings.i18n.languages as string[]);
+      return this.formatter(
+        value,
+        this.props.bindings.i18n.languages as string[]
+      );
     } catch (error) {
-      this.bindings.engine.logger.error(
+      this.props.bindings.engine.logger.error(
         `atomic-numeric-facet facet value "${value}" could not be formatted correctly.`,
         error
       );
@@ -276,8 +251,8 @@ export class NumericFacetCommon {
   }
 
   private areRangesEqual(
-    firstRange: InsightNumericRangeRequest | NumericRangeRequest,
-    secondRange: InsightNumericRangeRequest | NumericRangeRequest
+    firstRange: NumericRangeRequest,
+    secondRange: NumericRangeRequest
   ) {
     return (
       firstRange.start === secondRange.start &&
@@ -286,20 +261,17 @@ export class NumericFacetCommon {
     );
   }
 
-  private renderValue(
-    facetValue: InsightNumericFacetValue,
-    onClick: () => void
-  ) {
+  private renderValue(facetValue: NumericFacetValue, onClick: () => void) {
     const displayValue = this.formatFacetValue(facetValue);
     const isSelected = facetValue.state === 'selected';
-    switch (this.displayValuesAs) {
+    switch (this.props.displayValuesAs) {
       case 'checkbox':
         return (
           <FacetValueCheckbox
             displayValue={displayValue}
             numberOfResults={facetValue.numberOfResults}
             isSelected={isSelected}
-            i18n={this.bindings.i18n}
+            i18n={this.props.bindings.i18n}
             onClick={onClick}
           >
             <FacetValueLabelHighlight
@@ -314,7 +286,7 @@ export class NumericFacetCommon {
             displayValue={displayValue}
             numberOfResults={facetValue.numberOfResults}
             isSelected={isSelected}
-            i18n={this.bindings.i18n}
+            i18n={this.props.bindings.i18n}
             onClick={onClick}
           >
             <FacetValueLabelHighlight
@@ -328,7 +300,10 @@ export class NumericFacetCommon {
 
   private renderValuesContainer(children: VNode[]) {
     return (
-      <FacetValuesGroup i18n={this.bindings.i18n} label={this.label}>
+      <FacetValuesGroup
+        i18n={this.props.bindings.i18n}
+        label={this.props.label}
+      >
         <ul class="mt-3" part="values">
           {children}
         </ul>
@@ -347,7 +322,7 @@ export class NumericFacetCommon {
     return this.renderValuesContainer(
       this.valuesToRender.map((value) =>
         this.renderValue(value, () =>
-          this.displayValuesAs === 'link'
+          this.props.displayValuesAs === 'link'
             ? this.facetForRange!.toggleSingleSelect(value)
             : this.facetForRange!.toggleSelect(value)
         )
@@ -358,9 +333,9 @@ export class NumericFacetCommon {
   private renderNumberInput() {
     return (
       <atomic-facet-number-input
-        type={this.withInput!}
-        bindings={this.bindings}
-        label={this.label}
+        type={this.props.withInput!}
+        bindings={this.props.bindings}
+        label={this.props.label}
         filter={this.filter!}
         filterState={this.filter!.state}
       ></atomic-facet-number-input>
@@ -374,8 +349,8 @@ export class NumericFacetCommon {
   ) {
     return (
       <FacetHeader
-        i18n={this.bindings.i18n}
-        label={this.label}
+        i18n={this.props.bindings.i18n}
+        label={this.props.label}
         onClearFilters={() => {
           headerFocus.focusAfterSearch();
           if (this.filter?.state.range) {
@@ -386,7 +361,7 @@ export class NumericFacetCommon {
         }}
         numberOfSelectedValues={this.numberOfSelectedValues}
         isCollapsed={isCollapsed}
-        headingLevel={this.headingLevel}
+        headingLevel={this.props.headingLevel}
         onToggleCollapse={onToggleCollapse}
         headerRef={headerFocus.setTarget}
       ></FacetHeader>
@@ -407,7 +382,7 @@ export class NumericFacetCommon {
     if (!firstSearchExecuted) {
       return (
         <FacetPlaceholder
-          numberOfValues={this.numberOfValues}
+          numberOfValues={this.props.numberOfValues}
           isCollapsed={isCollapsed}
         ></FacetPlaceholder>
       );