+++ packages/atomic/src/components/common/facets/timeframe-facet-common.tsx
@@ -1,28 +1,7 @@
 import {h, VNode} from '@stencil/core';
-import {
-  DateFacet,
-  DateFacetValue,
-  DateFilter,
-  DateRangeRequest,
-  FacetConditionsManager,
-  RelativeDate,
-  SearchStatusState,
-} from '@coveo/headless';
 import dayjs from 'dayjs';
 import {FocusTargetController} from '../../../utils/accessibility-utils';
 import {getFieldValueCaption} from '../../../utils/field-utils';
-import {
-  InsightDateFacet,
-  InsightDateFacetValue,
-  InsightDateFilter,
-  InsightDateRangeOptions,
-  InsightDateRangeRequest,
-  InsightFacetConditionsManager,
-  InsightRelativeDate,
-  InsightSearchStatusState,
-} from '../../insight';
-import {InsightBindings} from '../../insight/atomic-insight-interface/atomic-insight-interface';
-import {Bindings} from '../../search/atomic-search-interface/atomic-search-interface';
 import {Hidden} from '../hidden';
 import {
   shouldDisplayInputForFacetRange,
@@ -34,38 +13,47 @@ import {FacetValueLink} from './facet-value-link/facet-value-link';
 import {FacetValueLabelHighlight} from './facet-value-label-highlight/facet-value-label-highlight';
 import {FacetValuesGroup} from './facet-values-group/facet-values-group';
 import {FacetHeader} from './facet-header/facet-header';
+import {AnyBindings} from '../interface/bindings';
+import {
+  DateFacet,
+  DateFacetValue,
+  DateFilter,
+  DateRangeOptions,
+  DateRangeRequest,
+  FacetConditionsManager,
+  RelativeDate,
+  RelativeDatePeriod,
+  RelativeDateUnit,
+  SearchStatusState,
+} from '../types';
 import {FacetInfo} from './facet-common-store';
 import {initializePopover} from '../../search/facets/atomic-popover/popover-type';
+import {randomID} from '../../../utils/utils';
 
-export interface Timeframe extends RelativeDate {
-  label?: string;
-}
-
-export interface InsightTimeframe extends InsightRelativeDate {
+export interface Timeframe {
+  period: RelativeDatePeriod;
+  unit?: RelativeDateUnit;
+  amount?: number;
   label?: string;
 }
 
 interface TimeframeFacetCommonOptions {
+  facetId?: string;
   host: HTMLElement;
-  bindings: Bindings | InsightBindings;
+  bindings: AnyBindings;
   label: string;
   field: string;
   headingLevel: number;
   dependsOn: Record<string, string>;
   withDatePicker: boolean;
-  getSearchStatusState(): InsightSearchStatusState | SearchStatusState;
-  buildDependenciesManager():
-    | InsightFacetConditionsManager
-    | FacetConditionsManager;
-  deserializeRelativeDate(date: string): InsightRelativeDate | RelativeDate;
-  buildDateRange(
-    config: InsightDateRangeOptions | InsightDateRangeOptions
-  ): InsightDateRangeRequest | DateRangeRequest;
-  initializeFacetForDatePicker(): InsightDateFacet | DateFacet;
-  initializeFacetForDateRange(
-    values: InsightDateRangeRequest[] | DateRangeRequest[]
-  ): InsightDateFacet | DateFacet;
-  initializeFilter(): InsightDateFilter | DateFilter;
+  setFacetId(id: string): string;
+  getSearchStatusState(): SearchStatusState;
+  buildDependenciesManager(): FacetConditionsManager;
+  deserializeRelativeDate(date: string): RelativeDate;
+  buildDateRange(config: DateRangeOptions): DateRangeRequest;
+  initializeFacetForDatePicker(): DateFacet;
+  initializeFacetForDateRange(values: DateRangeRequest[]): DateFacet;
+  initializeFilter(): DateFilter;
 }
 
 interface TimeframeFacetCommonRenderProps {
@@ -77,45 +65,17 @@ interface TimeframeFacetCommonRenderProps {
 }
 
 export class TimeframeFacetCommon {
-  private host: HTMLElement;
-  private bindings: Bindings | InsightBindings;
-  private label: string;
-  private field: string;
-  private headingLevel: number;
-  private dependsOn: Record<string, string>;
-  private withDatePicker: boolean;
   private facetId?: string;
-  private facetForDatePicker?: InsightDateFacet | DateFacet;
-  private facetForDateRange?: InsightDateFacet | DateFacet;
-  private filter?: InsightDateFilter | DateFilter;
-  private manualTimeframes: Timeframe[] | InsightTimeframe[] = [];
-  private dependenciesManager?:
-    | InsightFacetConditionsManager
-    | FacetConditionsManager;
-  private deserializeRelativeDate: (
-    date: string
-  ) => InsightRelativeDate | RelativeDate;
-  private getSearchStatusState: () =>
-    | InsightSearchStatusState
-    | SearchStatusState;
+  private facetForDatePicker?: DateFacet;
+  private facetForDateRange?: DateFacet;
+  private filter?: DateFilter;
+  private manualTimeframes: Timeframe[] = [];
+  private dependenciesManager?: FacetConditionsManager;
 
-  private buildDateRange: (
-    config: InsightDateRangeOptions | InsightDateRangeOptions
-  ) => InsightDateRangeRequest | DateRangeRequest;
-
-  constructor(props: TimeframeFacetCommonOptions) {
-    this.host = props.host;
-    this.bindings = props.bindings;
-    this.label = props.label;
-    this.field = props.field;
-    this.headingLevel = props.headingLevel;
-    this.dependsOn = props.dependsOn;
-    this.withDatePicker = props.withDatePicker;
-    this.deserializeRelativeDate = props.deserializeRelativeDate;
-    this.getSearchStatusState = props.getSearchStatusState;
-    this.buildDateRange = props.buildDateRange;
-
+  constructor(private props: TimeframeFacetCommonOptions) {
     this.validateProps();
+    this.facetId = this.determineFacetId;
+    this.props.setFacetId(this.facetId);
 
     this.manualTimeframes = this.getManualTimeframes();
 
@@ -125,23 +85,32 @@ export class TimeframeFacetCommon {
     // A second facet is initialized only to verify the results count. It is never used to display results to end user.
     // It serves as a way to determine if the input should be rendered or not, independent of the ranges configured in the component
     if (this.manualTimeframes.length > 0) {
-      this.facetForDateRange = props.initializeFacetForDateRange(
+      this.facetForDateRange = this.props.initializeFacetForDateRange(
         this.currentValues
       );
-      this.facetId = this.facetForDateRange.state.facetId;
     }
-    if (this.withDatePicker) {
-      this.facetForDatePicker = props.initializeFacetForDatePicker();
-      this.filter = props.initializeFilter();
 
-      if (!this.facetId) {
-        this.facetId = this.filter.state.facetId;
-      }
+    if (this.props.withDatePicker) {
+      this.facetForDatePicker = this.props.initializeFacetForDatePicker();
+      this.filter = this.props.initializeFilter();
     }
+
     if (this.facetForDateRange || this.filter) {
-      this.dependenciesManager = props.buildDependenciesManager();
+      this.dependenciesManager = this.props.buildDependenciesManager();
     }
     this.registerFacetToStore();
+  }
+
+  private get determineFacetId() {
+    if (this.props.facetId) {
+      return this.props.facetId;
+    }
+
+    if (this.props.bindings.store.get('dateFacets')[this.props.field]) {
+      return randomID(`${this.props.field}_`);
+    }
+
+    return this.props.field;
   }
 
   private get enabled() {
@@ -170,9 +139,9 @@ export class TimeframeFacetCommon {
 
   private get shouldRenderInput() {
     return shouldDisplayInputForFacetRange({
-      hasInput: this.withDatePicker,
+      hasInput: this.props.withDatePicker,
       hasInputRange: this.hasInputRange,
-      searchStatusState: this.getSearchStatusState(),
+      searchStatusState: this.props.getSearchStatusState(),
       facetValues: this.facetForDatePicker?.state?.values || [],
     });
   }
@@ -201,14 +170,14 @@ export class TimeframeFacetCommon {
     return !!this.filter?.state.range;
   }
 
-  public get currentValues(): InsightDateRangeRequest[] | DateRangeRequest[] {
+  public get currentValues(): DateRangeRequest[] {
     return this.manualTimeframes.map(({period, amount, unit}) =>
       period === 'past'
-        ? this.buildDateRange({
+        ? this.props.buildDateRange({
             start: {period, unit, amount},
             end: {period: 'now'},
           })
-        : this.buildDateRange({
+        : this.props.buildDateRange({
             start: {period: 'now'},
             end: {period, unit, amount},
           })
@@ -216,46 +185,42 @@ export class TimeframeFacetCommon {
   }
 
   public disconnectedCallback() {
-    if (this.host.isConnected) {
+    if (this.props.host.isConnected) {
       return;
     }
     this.dependenciesManager?.stopWatching();
   }
 
   private validateProps() {
-    validateDependsOn(this.dependsOn);
+    validateDependsOn(this.props.dependsOn);
   }
 
   private registerFacetToStore() {
-    if (!this.facetForDateRange) {
-      return;
-    }
-
     const facetInfo: FacetInfo = {
-      label: this.label,
+      label: this.props.label,
       facetId: this.facetId!,
-      element: this.host,
+      element: this.props.host,
     };
 
-    this.bindings.store.registerFacet('dateFacets', {
+    this.props.bindings.store.registerFacet('dateFacets', {
       ...facetInfo,
       format: (value) => this.formatFacetValue(value),
     });
 
-    initializePopover(this.host, {
+    initializePopover(this.props.host, {
       ...facetInfo,
       hasValues: () => this.hasValues,
       numberOfSelectedValues: () => this.numberOfSelectedValues,
     });
 
     if (this.filter) {
-      this.bindings.store.state.dateFacets[this.filter.state.facetId] =
-        this.bindings.store.state.dateFacets[this.facetId!];
+      this.props.bindings.store.state.dateFacets[this.filter.state.facetId] =
+        this.props.bindings.store.state.dateFacets[this.facetId!];
     }
   }
 
   private getManualTimeframes(): Timeframe[] {
-    return Array.from(this.host.querySelectorAll('atomic-timeframe')).map(
+    return Array.from(this.props.host.querySelectorAll('atomic-timeframe')).map(
       ({label, amount, unit, period}) => ({
         label,
         amount,
@@ -265,13 +230,13 @@ export class TimeframeFacetCommon {
     );
   }
 
-  private formatFacetValue(facetValue: InsightDateFacetValue | DateFacetValue) {
+  private formatFacetValue(facetValue: DateFacetValue) {
     try {
-      const startDate = this.deserializeRelativeDate(facetValue.start);
+      const startDate = this.props.deserializeRelativeDate(facetValue.start);
       const relativeDate =
         startDate.period === 'past'
           ? startDate
-          : this.deserializeRelativeDate(facetValue.end);
+          : this.props.deserializeRelativeDate(facetValue.end);
       const timeframe = this.getManualTimeframes().find(
         (timeframe) =>
           timeframe.period === relativeDate.period &&
@@ -281,19 +246,19 @@ export class TimeframeFacetCommon {
 
       if (timeframe?.label) {
         return getFieldValueCaption(
-          this.field,
+          this.props.field,
           timeframe.label,
-          this.bindings.i18n
+          this.props.bindings.i18n
         );
       }
-      return this.bindings.i18n.t(
+      return this.props.bindings.i18n.t(
         `${relativeDate.period}-${relativeDate.unit}`,
         {
           count: relativeDate.amount,
         }
       );
     } catch (error) {
-      return this.bindings.i18n.t('to', {
+      return this.props.bindings.i18n.t('to', {
         start: dayjs(facetValue.start).format('YYYY-MM-DD'),
         end: dayjs(facetValue.end).format('YYYY-MM-DD'),
       });
@@ -304,7 +269,7 @@ export class TimeframeFacetCommon {
       this.valuesToRender.map((value) => this.renderValue(value))
     );
   }
-  private renderValue(facetValue: InsightDateFacetValue) {
+  private renderValue(facetValue: DateFacetValue) {
     const displayValue = this.formatFacetValue(facetValue);
     const isSelected = facetValue.state === 'selected';
     return (
@@ -312,7 +277,7 @@ export class TimeframeFacetCommon {
         displayValue={displayValue}
         isSelected={isSelected}
         numberOfResults={facetValue.numberOfResults}
-        i18n={this.bindings.i18n}
+        i18n={this.props.bindings.i18n}
         onClick={() => this.facetForDateRange!.toggleSingleSelect(facetValue)}
       >
         <FacetValueLabelHighlight
@@ -325,7 +290,10 @@ export class TimeframeFacetCommon {
 
   private renderValuesContainer(children: VNode[]) {
     return (
-      <FacetValuesGroup i18n={this.bindings.i18n} label={this.label}>
+      <FacetValuesGroup
+        i18n={this.props.bindings.i18n}
+        label={this.props.label}
+      >
         <ul class="mt-3" part="values">
           {children}
         </ul>
@@ -340,8 +308,8 @@ export class TimeframeFacetCommon {
   ) {
     return (
       <FacetHeader
-        i18n={this.bindings.i18n}
-        label={this.label}
+        i18n={this.props.bindings.i18n}
+        label={this.props.label}
         onClearFilters={() => {
           headerFocus.focusAfterSearch();
           if (this.filter?.state.range) {
@@ -352,7 +320,7 @@ export class TimeframeFacetCommon {
         }}
         numberOfSelectedValues={this.numberOfSelectedValues}
         isCollapsed={isCollapsed}
-        headingLevel={this.headingLevel}
+        headingLevel={this.props.headingLevel}
         onToggleCollapse={onToggleCollapse}
         headerRef={headerFocus.setTarget}
       ></FacetHeader>
@@ -362,8 +330,8 @@ export class TimeframeFacetCommon {
   private renderDateInput() {
     return (
       <atomic-facet-date-input
-        bindings={this.bindings}
-        label={this.label}
+        bindings={this.props.bindings}
+        label={this.props.label}
         filter={this.filter!}
         filterState={this.filter!.state!}
       ></atomic-facet-date-input>