+++ packages/atomic/src/components/common/types.ts
@@ -1,17 +1,42 @@
-export type SearchStatus = {
-  state: {
-    hasResults: boolean;
-  };
-};
-
-export type Pager = {
-  isCurrentPage(page: number): boolean;
-  selectPage(page: number): void;
-  previousPage(): void;
-  nextPage(): void;
-  state: {
-    hasNextPage: boolean;
-    hasPreviousPage: boolean;
-    currentPages: number[];
-  };
-};
+// TODO: revisit using Core Headless Interfaces when using other sub-packages
+export {
+  SearchStatus,
+  SearchStatusState,
+  FacetSortCriterion,
+  SpecificFacetSearchResult,
+  FacetSearchState,
+  FacetState,
+  Facet,
+  CategoryFacetSearchResult,
+  CategoryFacetState,
+  CategoryFacetSearch,
+  FacetManagerPayload,
+  FacetValue,
+  FacetValueState,
+  FacetManager,
+  FacetConditionsManager,
+  FacetValueRequest,
+  CategoryFacetValueRequest,
+  Pager,
+  RelativeDatePeriod,
+  RelativeDateUnit,
+  RelativeDate,
+  DateFacetState,
+  DateFacet,
+  DateFacetValue,
+  DateRangeOptions,
+  DateRangeRequest,
+  DateFilterState,
+  DateFilter,
+  QuerySummary,
+  NumericFacetValue,
+  NumericFacetState,
+  NumericFacet,
+  NumericFilterState,
+  NumericFilter,
+  NumericRangeOptions,
+  NumericRangeRequest,
+  CategoryFacet,
+  CategoryFacetSortCriterion,
+  RangeFacetSortCriterion,
+} from '@coveo/headless';